Improve calling of external commands:
If the command isn't found at the path specified by the config.php file, then an error is reported.

For most commands, the error output is now correctly captured and displayed to the user.  In particular, this should help with issue 94, whereby an incorrectly configured diff would generate a null output with no further explanation.