Decouple MultiViews glue code from its initial configuration
The current wsvn.php requires the user to set some variables, if necessary, yet
it contains a huge block of non-modifiable code. If wsvn.php is changed in Git
the use won't notice or even worse, the package maintainer will overwrite the
newer version with the user's one.

This change decouples the glue code from the MultiViewsIndex making it actually
configurable w/o losing updates.