Revert "Fix access control with a subpath repository", because it seems to difficult to me.
configclass::Repository already contains the subpath and can use it as needed internally, no need to force users to check if a subpath is available or not and build paths on their own. In rss.php line 60, SVNRepository has been created using $rep, which contains the subpath itself, to query that subpath by accessing $rep within SVNRepository.

In lack of a working setup I'm unable to test this and want to provide this commit to @Tom-H- to check.

This reverts commit 456c7f49c3d109fa9b2f0fff8da8040a8d643742.