Try to support one WebSVN-installation hosting multiple different SVNParentPaths.
"svnserve" is able to host SVN-repos in subdirs of a given root path, mod_dav_svn and WebSVN are not and need to get each subdir configured manually. That's especially difficult for WebSVN because it includes a config file based on a path by convention always, which is relative to the installation of WebSVN itself. So in case one wanted to use different root paths to SVN-repos, one would need multiple installations of WebSVN to provide different config files with those different SVN root paths.

This patch changes that to an approach getting the config file to use from some environment variable which can be set by the web server based on the requested location and stuff. This way one installation of WebSVN can serve different locations using different config files.