Support one WebSVN hosting multiple different parent paths for repos. (#87)
"svnserve" is able to host SVN repos in subdirs of a given root path, mod_dav_svn and WebSVN are not and need to get each subdir configured manually. That's especially difficult for WebSVN because it includes a config file based on a path by convention which is relative to the installation of WebSVN itself. So in case one wanted to use different parent paths to repos, one would need multiple installations of WebSVN to provide different config files with those different root paths.

This patch changes that to an approach optionally and additionally getting the config file to use from some environment variable which can be set by the web server based on the requested location and stuff. This way one installation of WebSVN can serve different locations using different config files because the default config required by path and the one provided by the environment variable are layered:

"includes/config.php" is included always if it exists, exactly like with the original implementation. If an additional path is given using the special environment variable, that is required additionally, allowing overriding "config.php" or adding things not available in the former file at all. This allows configuring one installation of WebSVN using one "config.php" regarding templates and common stuff, while the file given by the environment variable might specify different paths to actual repos to show only.