Deprecated use of "count" in "levenshtein2".
"levenshtein2" implements the algorithm on its own in case of too large given strings based on words. The function uses "count" for this, which is deprecated on strings and should be taking words instead of characters into account anyway. So indexes are now calculated based on the words.

This fixes #95.