Show files in subdirs with parents with restricted access.
Calling "showDirFiles" with a level of "0" always leads to possible failed access checks when "/", "/tags" etc. are NOT visible to the current user, while "/tags/DIR1" might be. This is because of early aborts when checking access starting from the root of the repo.

Fixes https://github.com/websvnphp/websvn/issues/146.