Make the root-phrase of the breadcrumb path configurable. (#152)
The approach is to 1. get rid of the hard-coded string and make it a language property instead, so one can have phrases like "(root)" in English, "(wurzel)" in German or some arbitrary other string easily.

Second, a new config is introduced to let WebSVN calculate the name of that root-phrase based on the name of the current repo automatically. that might make sense for a lot of people and is what's actually requested in the linked issue by @didiez.

Third, the breadcrumb path rendered into templates has been splitted into root and non-root, so that templates can decide to use an arbitrary different root or none at all. Before, templates only received the whole rendered breadcrumb, which isn't pretty flexible. Things can easily be reverted with mass-replace anyway.

Fixes https://github.com/websvnphp/websvn/issues/150.